Enable JAX float64 precision for numerical stability
Add jax_enable_x64 configuration at package init and upgrade remaining
float32 casts to float64 in Kalman update masks, Rao-Blackwell weights,
SVI losses, and missing-data preprocessing. The bulk of the float64
migration landed in the prior commit alongside the import rewrites.

+++ apps/data-pipeline/src/causal_ssm_agent/models/ssm/inference/targets/base.py
@@ -143,7 +143,7 @@ def preprocess_missing_data(
 
     # Build per-timestep R with inflated variance for missing entries
     T, n_manifest = observations.shape
-    mask_float = obs_mask.astype(jnp.float32)  # (T, n_manifest)
+    mask_float = obs_mask.astype(jnp.float64)  # (T, n_manifest)
     # (T, n_manifest) diagonal inflation values
     inflation = (1.0 - mask_float) * MISSING_DATA_LARGE_VAR
     # Broadcast R to (T, n_manifest, n_manifest) and add diagonal inflation
